extracted story mock helper and additional clean up
