+++ lib/ItemRow/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import PropTypes from 'prop-types';
+import { node, string, bool, oneOfType, arrayOf } from 'prop-types';
 import cx from 'classnames';
 import { ItemRowName } from './ItemRowName';
 import { ItemRowAside } from './ItemRowAside';
@@ -19,13 +19,10 @@ const ItemRow = ({ children, stacked, className, bordered, ...rest }) => {
 };
 
 ItemRow.propTypes = {
-  children: PropTypes.oneOfType([
-    PropTypes.node,
-    PropTypes.arrayOf(PropTypes.node)
-  ]).isRequired,
-  stacked: PropTypes.bool,
-  bordered: PropTypes.bool,
-  className: PropTypes.string
+  children: oneOfType([node, arrayOf(node)]).isRequired,
+  stacked: bool,
+  bordered: bool,
+  className: string
 };
 
 ItemRow.defaultProps = {
