+++ lib/ItemRow/stories/ItemRowStory.js
@@ -2,17 +2,9 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { boolean, object, text } from '@storybook/addon-knobs';
 import { ItemRow, StatusIndicator } from '../../index';
-import AvatarGroup from '../../Avatar/AvatarGroup';
-import Avatar from '../../Avatar';
 import notes from '../README.md';
 import Icon from '../../Icon';
-
-const createAvatarGroup = () => (
-  <AvatarGroup>
-    <Avatar smallSize initials="RS" name="Richard Swagshaw" bordered />
-    <Avatar smallSize initials="JD" name="James Darracott" bordered />
-  </AvatarGroup>
-);
+import { createAvatarGroup } from '../../Avatar/stories/createAvatarGroup';
 
 const createStatusIndicator = props => <StatusIndicator {...props} />;
